rename: svg 파일명 형식에 맞게 수정
